requirements.txt
numpy>=1.20.0
pandas>=1.3.0
torch>=1.10.0
scikit-learn>=1.0.0
scipy>=1.7.0

matplotlib>=3.4.0
altair>=4.0.0

requests>=2.25.0
beautifulsoup4>=4.9.0
ripe.atlas.cousteau>=1.4.0
py-radix>=0.10.0

tqdm>=4.60.0
loguru>=0.5.0

ijson>=3.1.0
msgpack>=1.0.0
